backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.10
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==1.10.9
python-multipart==0.0.6